Split MediaWikiServicesHook
